chore(ts): migrate App component to TypeScript
